hrms/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
colorama==0.4.6
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.4.0
pycparser==2.21
PySocks==1.7.1
pytest==8.0.2
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
